pyproject.toml
[project]
name = "blog"
version = "0.1.0"
description = "Ventanita Argentina 🇦🇷"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "django-allauth[socialaccount]>=65.14.3",
    "django-countries>=8.2.0",
    "mercadopago>=2.3.0",
    "psycopg2-binary>=2.9.11",
    "python-dotenv>=1.2.1",
    "reportlab>=4.4.10",
    "wagtail>=7.3",
    "wagtail-localize>=1.13",
    "zeep>=4.3.2",
]

[dependency-groups]
dev = [
    "django-debug-toolbar>=6.2.0",
    "django-extensions>=4.1",
]

[tool.ruff]
line-length=79
exclude=["*migrations*", "static"]

[tool.ruff.lint.isort]
combine-as-imports=true
force-single-line=false
known-first-party=["blog", "base", "home", "search", "users", "partners", "orders", "payments"]
known-third-party=["wagtail", "rest_framework"]
lines-after-imports=2
section-order=["future", "standard-library", "django", "wagtail", "third-party", "first-party", "local-folder"]

[tool.ruff.lint.isort.sections]
"django"=["django"]
"wagtail"=["wagtail"]

[tool.djlint]
blank_line_before_tag="load,extends"
blank_line_after_tag="load,extends,endblock"
close_void_tags=true
custom_blocks="toc,example"
custom_html="mjml"
exclude=".venv,venv,.tox,.eggs,.git"
extend_exclude=".custom"
format_attribute_template_tags=true
format_css=true
format_js=true
ignore_blocks="raw,example"
indent=2
max_attribute_length=120
max_line_length=120
profile="django"
# preserve_blank_lines=true
use_gitignore=true

[tool.djlint.css]
indent_size=5

[tool.djlint.js]
indent_size=5

[tool.coverage.run]
branch = true
omit = [
    # omit anything in a migrations directory anywhere
    # "*/migrations/*",
    # omit the tests themselfves
    # "*/tests/*",
    # omit these single files
    # "main/asgi.py",
    # "main/wsgi.py",
    # "main/settings.py",
    # "manage.py",
    # "*/urls.py",
    # "*/__init__.py",
]

[tool.coverage.report]
show_missing = true
skip_covered = false

[tool.coverage.html]
show_contexts = true
